feat: implement token bucket rate limiter and refactor server integration
- Introduced a new TokenBucketLimiter in the pkg package to manage rate limiting for tools.
- Updated the current_time_server and everything examples to utilize the new rate limiter.
- Refactored server middleware to accept the new pkg.RateLimiter interface, enhancing flexibility and maintainability.
- Removed deprecated components references in favor of the new implementation.
